requirements.txt
streamlit
openai
langchain
python-docx
nltk
language-tool-python
